Refactor deployed environment CI checks
The updates for #66 are substantially changing the way the
deployed environment checks work in the test suite.

To make that change easier to review, first deduplicate
the deployed environment checking code.
